Move to directory; add sanitize unit test
