+++ src/applications/static-pages/homepage-banner/HomepageBanner.jsx
