+++ src/applications/static-pages/homepage-banner/createHomepageBanner.js
